+++ src/extractor/compressed_edge_container.cpp
@@ -2,10 +2,11 @@
 #include "util/log.hpp"
 
 #include <boost/assert.hpp>
-#include <boost/filesystem.hpp>
-#include <boost/filesystem/fstream.hpp>
 #include <boost/numeric/conversion/cast.hpp>
 
+#include <filesystem>
+#include <fstream>
+#include <iostream>
 #include <limits>
 #include <string>